Updates related to final report
